fix(.\controllers\mainPage.js; routes.js): Контроллер главной страницы
